refactor: extract reusable note preview source from NoteEmbedWidget
Move the note content reader (openTab -> cache -> async fetch) plus
renderEmbedPreview / extractSection / appendFormattedInline out of
livePreviewBlocks.ts into a shared notePreviewSource.ts module.

NoteEmbedWidget now consumes findPreviewNote + getNotePreviewContentState
instead of owning the lookup and cache logic inline. This establishes the
single source of truth that the upcoming wikilink hover preview will reuse,
so embeds and hover share one reader/renderer instead of duplicating it.

No behavior change. invalidateLivePreviewNoteCache is re-exported under its
historical name for existing callers.
